Assignment 4 - First commit
Most data is stored in databases, for ready access and organization.  Our course data is backed up by IT in databases which makes our data easy to access and use.

Write a program that creates a database of courses.  It will either read from a file of courses, or allow the user to add one course at a time.